--- backend/driver/amazons3/amazons3_test.go
@@ -1,23 +1,0 @@
-package amazons3
-
-import (
-	"testing"
-
-	"github.com/containerops/wrench/setting"
-)
-
-func Test_amazons3save(t *testing.T) {
-	var err error
-	var url string
-
-	if err = setting.SetConfig("../../../conf/containerops.conf"); err != nil {
-		t.Error(err)
-	}
-
-	file := "amazons3_test.go"
-	url, err = amazons3save(file)
-	if err != nil {
-		t.Error(err)
-	}
-	t.Log(url)
-}
