+++ test/backend/amazons3_test.go
@@ -1,13 +1,13 @@
-package qiniu
+package backend
 
 import (
-	"net/http"
 	"testing"
 
+	"github.com/containerops/dockyard/backend/driver/amazons3"
 	"github.com/containerops/wrench/setting"
 )
 
-func Test_qiniusave(t *testing.T) {
+func Test_amazons3save(t *testing.T) {
 	var err error
 	var url string
 
@@ -15,13 +15,10 @@ func Test_qiniusave(t *testing.T) {
 		t.Error(err)
 	}
 
-	file := "qiniu_test.go"
-	url, err = qiniusave(file)
+	file := "amazons3_test.go"
+	url, err = amazons3save(file)
 	if err != nil {
 		t.Error(err)
 	}
-	_, err = http.Get(url)
-	if err != nil {
-		t.Error(err)
-	}
+	t.Log(url)
 }
